+++ api.Tests/ApiRoutesTests/ConstitutionArticleApiIntegrationTests.cs
@@ -4,13 +4,14 @@ using api.Utils;
 
 namespace api.Tests.ApiRoutesTests;
 
-public class ConstitutionArticleApiIntegrationTests : IClassFixture<CustomWebApplicationFactory>, IDisposable
+public class ConstitutionArticleApiIntegrationTests : IClassFixture<CustomWebApplicationFactory>
 {
-    private readonly HttpClient _client = new CustomWebApplicationFactory().CreateClient();
+    private readonly HttpClient _client;
 
-    public void Dispose()
+    public ConstitutionArticleApiIntegrationTests(CustomWebApplicationFactory factory)
     {
-        _client.Dispose();
+        factory.ResetDatabase();
+        _client = factory.CreateClient();
     }