+++ api.Tests/ApiRoutesTests/CountryApiIntegrationTests.cs
@@ -3,10 +3,15 @@ using api.Models;
 
 namespace api.Tests.ApiRoutesTests;
 
-public class CountryApiIntegrationTests(CustomWebApplicationFactory factory)
-    : IClassFixture<CustomWebApplicationFactory>
+public class CountryApiIntegrationTests : IClassFixture<CustomWebApplicationFactory>
 {
-    private readonly HttpClient _client = factory.CreateClient();
+    private readonly HttpClient _client;
+
+    public CountryApiIntegrationTests(CustomWebApplicationFactory factory)
+    {
+        factory.ResetDatabase();
+        _client = factory.CreateClient();
+    }
 
     [Fact]
     public async Task GetCountry_ReturnsAllCountryData()