+++ api/Program.cs
@@ -1,11 +1,11 @@
 using System.Text.Json.Serialization;
 using api;
 using api.Routes;
-using Microsoft.OpenApi.Models;
 using Microsoft.AspNetCore.Http.Json;
 using api.Utils;
 using System.Net;
 using api.Const;
+using Microsoft.OpenApi;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -13,7 +13,7 @@ builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen(static options =>
 {
     options.EnableAnnotations();
-    options.SwaggerDoc("v1", new OpenApiInfo
+    options.SwaggerDoc("v1", new OpenApiInfo()
     {
         Version = VersionInfo.CurrentVersion,
         Title = "API Colombia",